+++ app/src/main/res/layout/activity_report_location.xml
@@ -5,7 +5,7 @@
     android:id="@+id/main"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".presentation.ui.report.ReportLocationActivity">
+    tools:context=".presentation.ui.report.dialog.ReportLocationActivity">
 
     <WebView
         android:layout_width="match_parent"
